requirements.txt
Flask==3.1.0
gTTS==2.5.4
openai==0.28.0
python-dotenv==1.0.1
